agent-driven/ccode-llm-logic-eval/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
python-constraint>=1.4.0
